working version -without auto launcher -with memory meta database -query manager getts data from mongo
