+++ pom.xml
@@ -4,7 +4,7 @@
 
     <modelVersion>4.0.0</modelVersion>
     <groupId>eu.squadd.spring-batch-test</groupId>
-    <artifactId>SpringBatchTest</artifactId>
+    <artifactId>eu.squadd.springbatchtest</artifactId>
     <version>1.0.0</version>
 
     <parent>
@@ -63,13 +63,9 @@
             <artifactId>cassandra-driver-mapping</artifactId>
         </dependency>
         <dependency>
-            <groupId>org.apache.derby</groupId>
-            <artifactId>derby</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.derby</groupId>
-            <artifactId>derbyclient</artifactId>
-            <version>10.13.1.1</version>
+            <groupId>com.h2database</groupId>
+            <artifactId>h2</artifactId>
+            <scope>runtime</scope>
         </dependency>
         <dependency>
             <groupId>org.springframework.boot</groupId>
@@ -103,6 +99,13 @@
             <version>3.0.8.RELEASE</version>
             <type>jar</type>
         </dependency>-->
+        <dependency>
+            <groupId>org.mongodb</groupId>
+            <artifactId>mongodb-driver</artifactId>
+            <version>3.4.3</version>
+            <type>jar</type>
+        </dependency>
+        
     </dependencies>
 
     <profiles>
@@ -185,7 +188,7 @@
         </profile>
     </profiles>
     
-    <name>SpringBatchTest</name>
+    <name>spring-batch-test</name>
 </project>
 
 <!--
