+++ src/main/java/eu/squadd/batch/domain/mongo/MongoUser.java
@@ -1,0 +1,27 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package eu.squadd.batch.domain.mongo;
+
+/**
+ *
+ * @author smoczyna
+ */
+public class MongoUser {
+
+    private String name;
+
+    public MongoUser(String name) {
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+}
