+++ src/main/java/eu/squadd/batch/listeners/GenericStepExecutionListener.java
@@ -10,8 +10,6 @@ import eu.squadd.batch.utils.WholesaleBookingProcessorHelper;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.batch.core.ExitStatus;
-import org.springframework.batch.core.ItemProcessListener;
-import org.springframework.batch.core.SkipListener;
 import org.springframework.batch.core.StepExecution;
 import org.springframework.batch.core.StepExecutionListener;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -20,23 +18,20 @@ import org.springframework.beans.factory.annotation.Autowired;
  *
  * @author smorcja
  */
-public class GenericStepExecutionListener implements StepExecutionListener, SkipListener, ItemProcessListener {
+public class GenericStepExecutionListener implements StepExecutionListener {
 
     @Autowired
     WholesaleBookingProcessorHelper processingHelper;
     
     private static final Logger LOGGER = LoggerFactory.getLogger(GenericStepExecutionListener.class);
-    private Long recordCount;
+    
     
     @Override
-    public void beforeStep(StepExecution se) { 
-        this.recordCount = 0L;
-        //this.processingHelper.setStepExecutionContext(se.getExecutionContext());
+    public void beforeStep(StepExecution se) {
     }
 
     @Override
     public ExitStatus afterStep(StepExecution se) {
-        //this.processingHelper.setStepExecutionContext(null);
         LOGGER.info(String.format(Constants.JOB_EXECUTION_FINISHED, se.getReadCount(), se.getWriteCount()));        
         LOGGER.info(String.format(Constants.WHOLESALE_REPORT_NO, this.processingHelper.getCounter(Constants.WHOLESALES_REPORT)));
         LOGGER.info(String.format(Constants.SUBLEDGER_REPORD_NO, this.processingHelper.getCounter(Constants.SUBLEDGER)));
@@ -48,31 +43,5 @@ public class GenericStepExecutionListener implements StepExecutionListener, Skip
         return se.getExitStatus();
     }
 
-    @Override
-    public void onSkipInRead(Throwable exception) {
-        LOGGER.error(String.format(Constants.READER_EXCEPTION, exception.getMessage()));
-    }
-
-    @Override
-    public void onSkipInWrite(Object s, Throwable exception) {
-        LOGGER.error(String.format(Constants.WRITER_EXCEPTION, exception.getMessage()));
-    }
-
-    @Override
-    public void onSkipInProcess(Object inputRecord, Throwable exception) {
-        LOGGER.error(inputRecord.toString());
-    }
-
-    @Override
-    public void beforeProcess(Object t) {
-        LOGGER.info(String.format(Constants.PROCESSING_RECORD, ++recordCount));
-    }
-
-    @Override
-    public void afterProcess(Object t, Object s) {
-    }
-
-    @Override
-    public void onProcessError(Object t, Exception excptn) {
-    }
+    
 }
