--- src/main/resources/data/waiting-room/bookdate.csv
@@ -1,1 +1,0 @@
-09/01/2017,09/30/2017,08/26/2017,09/25/2017,25
