--- src/main/resources/data/waiting-room/offset.csv
@@ -1,10 +1,0 @@
-0000000039,0000000191
-0000000040,0000000191
-0000000041,0000000187
-0000000043,0000000191
-0000000044,0000000190
-0000000045,0000000186
-0000000046,0000000194
-0000000048,0000000191
-0000000428,0000000187
-0000000429,0000000190
