--- src/main/resources/log4j2.xml
@@ -1,12 +1,0 @@
-<Configuration status="INFO">
-    <Appenders>
-        <Console name="console" target="SYSTEM_OUT">
-            <PatternLayout pattern="%msg%n"/>
-        </Console>
-    </Appenders>
-    <Loggers>
-        <Root level="info" additivity="false">
-            <AppenderRef ref="console"/>
-        </Root>
-    </Loggers>
-</Configuration>
