--- src/test/resources/cassandra.properties
@@ -1,10 +1,0 @@
-cassandra.FccCgsaMapEndDate=12/31/9999
-cassandra.FinancialMarketMapEndDate=12/31/9999
-cassandra.GLMarketLegalEntityEndDate=12/31/9999
-cassandra.GLMarketMapType=D
-cassandra.GLMarketEndDate=12/31/9999
-cassandra.AlternateBookingType=D
-cassandra.FinancialEventCategoryTable=financialeventcategory
-cassandra.FinancialMarketTable=financialmarket
-cassandra.ProductTable=product
-cassandra.DataEventTable=dataevent
